requirements/dev.txt
-r base.txt

django-debug-toolbar==1.2.2
sqlparse==0.1.14